fix: resolve review comments for get_team tool and integration tests
